fixing something

+++ src/database/migrations/20200607084034_incomes.js
@@ -2,7 +2,7 @@ exports.up = function (knex) {
   return knex.schema.createTable("incomes", (table) => {
     table.increments("id").primary();
     table.string("name").notNullable();
-    table.bigInteger("value").notNullable();
+    table.float("value").notNullable();
   });
 };
